Add files via upload
Task 2: Data Modeling

1


2


3


4


5

Here is your task
Step 1: Data modeling

Look at the data model provided in the additional resources. Look at all the data that is now available from the client and decide what you want to use for the modeling of the problem statement. This should take 10-15 minutes.

Step 2: Strategic planning

Come up with a plan as to how you’ll use this data to solve the problem statement that the client has positioned. This plan will be used to describe to the client how we are planning to complete the remaining work and to build trust with the client as a domain expert. If you need some guidance, use the provided resource video that describes the high-level overview of a data science project. This should take 10-15 minutes.

Step 3: Communication

Summarize the data that you want to make use of and the strategic plan of action in a single PowerPoint slide. This will be sent to the Data Science team leader and the client, so be sure to be concise and use business-friendly language. This should take 10-15 minutes.